fix race-condition locking up the firmware after instant-stop
When all block buffers are in use, some waiting loops store (block_buffer_head+1)
before the loop starts, under the assumption that no other code can add a block
to the buffer. Which is kind of correct. But if an instant-stop did happen during
such a loop it got stuck there forever.